requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
